+++ pypeit/scripts/ql_keck_mosfire.py
@@ -29,7 +29,8 @@ from pypeit import spec2dobj
 from pypeit import coadd2d
 from pypeit import specobjs
 from pypeit import slittrace
-from pypeit import reduce
+from pypeit import extraction
+from pypeit import find_objects
 from pypeit import calibrations
 from pypeit.display import display
 from pypeit.images import buildimage
@@ -110,18 +111,21 @@ def run_pair(A_files, B_files, caliBrate, spectrograph, det, parset, show=False,
     # Background Image?
     sciImg = sciImg.sub(buildimage.buildimage_fromlist(spectrograph, det, parset['scienceframe'], list(B_files), bpm=caliBrate.msbpm, slits=caliBrate.slits, ignore_saturation=False),
             parset['scienceframe']['process'])
-    # Instantiate Reduce object
+    # Instantiate FindObjects object
     # Required for pypeline specific object
     # At instantiaton, the fullmask in self.sciImg is modified
-    redux = reduce.Reduce.get_instance(sciImg, spectrograph, parset, caliBrate, 'science', bkg_redux=True, show=show,
-                                       det=det)
 
-    # skymodel, objmodel, ivarmodel, outmask, sobjs, scaleimg, waveimg, tilts = redux.run(
-    #     std_trace=std_trace, return_negative=True, show_peaks=show)
+    # DP: Should find_negative be True here?
+    objFind = find_objects.FindObjects(sciImg, spectrograph, parset, caliBrate, 'science', bkg_redux=True,
+                                       find_negative=True, show=show)
 
-    global_sky, sobjs_obj, skymask = redux.run_objfind(std_trace=std_trace, show_peaks=show)
-    skymodel, objmodel, ivarmodel, outmask, sobjs, scaleimg, waveimg, tilts = redux.run_extraction(
-        global_sky, sobjs_obj, skymask, return_negative=True)
+    global_sky, sobjs_obj = objFind.run(std_trace=std_trace, show_peaks=show)
+
+    # Instantiate Extract object
+    extract = extraction.Extract.get_instance(sciImg, sobjs_obj, spectrograph, parset, caliBrate,
+                                              'science', bkg_redux=True, find_negative=True, show=show)
+    skymodel, objmodel, ivarmodel, \
+    outmask, sobjs, scaleimg, waveimg, tilts = extract.run_extraction(global_sky, sobjs_obj , return_negative=True)
 
     # TODO -- Do this upstream
     # Tack on detector
@@ -131,7 +135,7 @@ def run_pair(A_files, B_files, caliBrate, spectrograph, det, parset, show=False,
     # Construct table of spectral flexure
     spec_flex_table = Table()
     spec_flex_table['spat_id'] = caliBrate.slits.spat_id
-    spec_flex_table['sci_spec_flexure'] = redux.slitshift
+    spec_flex_table['sci_spec_flexure'] = extract.slitshift
 
     # Construct the Spec2DObj with the positive image
     spec2DObj_A = spec2dobj.Spec2DObj(sciimg=sciImg.image,